fix: parse utils
