+++ apps/mobile/tests/unit/utils/format.test.ts
@@ -1,8 +1,6 @@
 import {
   formatAddress,
   formatDate,
-  formatLabel,
-  formatLabelTags,
   formatNumber,
   formatTime
 } from '@/utils/format'
@@ -53,35 +51,6 @@ describe('format utils', () => {
 
     it('should work with number date', () => {
       expect(formatDate(1711639918000)).toBe('Mar 28, 2024')
-    })
-  })
-
-  describe('formatLabel', () => {
-    it('should return label with no tags', () => {
-      const result = formatLabel('Test label')
-      expect(result.label).toBe('Test label')
-      expect(result.tags).toHaveLength(0)
-    })
-
-    it('should return label with tags', () => {
-      const result = formatLabel('Test label #kyc #satsigner')
-      expect(result.label).toBe('Test label')
-      expect(result.tags).toEqual(['kyc', 'satsigner'])
-    })
-  })
-
-  describe('formatLabelTags', () => {
-    it('should return only label', () => {
-      const result = formatLabelTags('My label', [])
-      expect(result).toBe('My label')
-    })
-    it('should return label and tags', () => {
-      const result = formatLabelTags('My label', ['endthefed', 'nokyc'])
-      expect(result).toBe('My label #endthefed #nokyc')
-    })
-    it('should return only tags', () => {
-      const result = formatLabelTags('', ['endthefed', 'nokyc'])
-      expect(result).toBe('#endthefed #nokyc')
     })
   })
 })
