+++ apps/mobile/utils/format.ts
@@ -45,22 +45,6 @@ function formatTimestamp(date: Date) {
   return Math.floor(date.getTime() / 1000)
 }
 
-function formatLabel(rawLabel: string) {
-  const matches = rawLabel.match(/#\w[\w\d]+/g)
-  if (!matches) return { label: rawLabel, tags: [] }
-
-  const tags = matches.map((match) => match.replace('#', ''))
-  const label = rawLabel.replace(/#.*/, '').trim()
-  return { label, tags }
-}
-
-function formatLabelTags(label: string, tags: string[]) {
-  const trimmedLabel = label.trim()
-  if (tags.length === 0) return trimmedLabel
-  const labelTagSeparator = label.length === 0 ? '' : ' '
-  return trimmedLabel + labelTagSeparator + tags.map((t) => '#' + t).join(' ')
-}
-
 function formatPageUrl(path: string, params: PageParams) {
   let url = '/' + (path || '')
 
@@ -107,8 +91,6 @@ export {
   formatConfirmations,
   formatDate,
   formatFiatPrice,
-  formatLabel,
-  formatLabelTags,
   formatNumber,
   formatPageUrl,
   formatPercentualChange,
